condensed audit code to one file
